+++ noodles-bcf/src/header/string_maps.rs
@@ -42,7 +42,7 @@ impl StringMaps {
     ///
     /// let header = vcf::Header::builder()
     ///     .add_info(info::Key::TotalDepth, Map::<Info>::from(info::Key::TotalDepth))
-    ///     .add_filter(Map::<Filter>::new("q10", "Quality below 10"))
+    ///     .add_filter("q10", Map::<Filter>::new("q10", "Quality below 10"))
     ///     .add_format(Map::<Format>::from(format::Key::ReadDepth))
     ///     .add_contig(Map::<Contig>::new("sq0".parse()?))
     ///     .build();
@@ -77,7 +77,7 @@ impl StringMaps {
     ///
     /// let header = vcf::Header::builder()
     ///     .add_info(info::Key::TotalDepth, Map::<Info>::from(info::Key::TotalDepth))
-    ///     .add_filter(Map::<Filter>::new("q10", "Quality below 10"))
+    ///     .add_filter("q10", Map::<Filter>::new("q10", "Quality below 10"))
     ///     .add_format(Map::<Format>::from(format::Key::ReadDepth))
     ///     .add_contig(Map::<Contig>::new("sq0".parse()?))
     ///     .build();
@@ -423,8 +423,8 @@ mod tests {
                 info::Key::TotalDepth,
                 Map::<Info>::from(info::Key::TotalDepth),
             )
-            .add_filter(Map::<Filter>::pass())
-            .add_filter(Map::<Filter>::new("q10", "Quality below 10"))
+            .add_filter("PASS", Map::<Filter>::pass())
+            .add_filter("q10", Map::<Filter>::new("q10", "Quality below 10"))
             .add_format(Map::<Format>::from(format::Key::Genotype))
             .add_format(Map::<Format>::from(format::Key::ReadDepth))
             .add_alternative_allele(Map::<AlternativeAllele>::new(