+++ noodles-vcf/src/header/parser.rs
@@ -122,7 +122,7 @@ fn parse_record(
     builder = match record {
         Record::FileFormat(_) => return Err(ParseError::UnexpectedFileFormat),
         Record::Info(info) => builder.add_info(info.id().clone(), info),
-        Record::Filter(filter) => builder.add_filter(filter),
+        Record::Filter(filter) => builder.add_filter(filter.id().clone(), filter),
         Record::Format(format) => builder.add_format(format),
         Record::AlternativeAllele(alternative_allele) => {
             builder.add_alternative_allele(alternative_allele)